Update namespaces for models that moved back to the Entities folder
