+++ Source/Application/Commands/CreateJobCommand.cs
@@ -1,5 +1,5 @@
 ﻿using Application.Common.Interfaces;
-using Domain;
+using Domain.Entities;
 using MediatR;
 using System;
 using System.Threading;
