+++ Source/Application/Common/Interfaces/IJobRepository.cs
@@ -1,5 +1,4 @@
-﻿using Domain;
-using System.Threading.Tasks;
+﻿using Domain.Entities;
 
 namespace Application.Common.Interfaces
 {
