+++ Source/Application/DTOs/CreateJobRequest.cs
@@ -1,5 +1,5 @@
 using System;
-using Domain;
+using Domain.Entities;
 
 namespace Application.DTOs
 {
